--- a/src/layers/generic.h
+++ b/src/layers/generic.h
@@ -147,6 +147,7 @@ private:
     template<typename T> Expr constant(const std::vector<T>& data) const { return constant(Shape{(int)data.size()}, data); } // same as constant() but assuming vector
     Expr indices(const std::vector<uint32_t>& data) const { return graph()->indices(data); } // actually the same as constant(data) for this data type
     std::vector<float> getFactorMasks(size_t factorGroup, const std::vector<WordIndex>& indices) const;
+    Expr getLemmaHasFactorGroupTensor() const;
     Expr addFactorMaxesHelper(Expr indices=nullptr) const;
 private:
     // members