… CHANGELOG.md
… src/CMakeLists.txt
… src/tensors/tensor_operators.h
… src/translator/beam_search.cpp